Moved new Stomp into network package and added a note
